# public/swagger-memory-shared/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "description": "statusがMEMBERのユーザ情報をすべて取得する",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::index",
                "parameters": [
                    {
                        "name": "name@like",
                        "in": "query",
                        "description": "ユーザ名の検索値(あいまい検索に対応)",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "sort_name",
                        "in": "query",
                        "description": "ユーザ名でソート",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "statusがMEMBERのユーザデータを返す",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/user_list"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error_message": {
                                            "description": "サーバエラー用のメッセージを表示",
                                            "type": "string",
                                            "example": "ユーザ情報を取得出来ませんでした"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ]
            },
            "post": {
                "tags": [
                    "users"
                ],
                "description": "ユーザデータを保存する",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "request": {
                                        "$ref": "#/components/schemas/user_register"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "info_message": {
                                            "description": "保存成功のメッセージを表示",
                                            "type": "string",
                                            "example": "ユーザ情報を登録しました"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error_message": {
                                            "description": "サーバエラー用のメッセージを表示",
                                            "type": "string",
                                            "example": "ユーザ情報の登録に失敗しました"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ]
            }
        },
        "/api/users/{user}": {
            "get": {
                "tags": [
                    "users"
                ],
                "description": "指定したユーザの情報をすべて取得する",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::show",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ユーザ名",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/user_list"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "存在しないユーザのページをリクエストした場合",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error_message": {
                                            "description": "検索結果が0件であることを表すメッセージを表示",
                                            "type": "string",
                                            "example": "指定したユーザは存在しません"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error_message": {
                                            "description": "サーバエラー用のメッセージを表示",
                                            "type": "string",
                                            "example": "ユーザ情報を取得出来ませんでした"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "user_list": {
                "required": [
                    "id",
                    "name",
                    "email",
                    "status",
                    "image_file"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "test1"
                    },
                    "email": {
                        "type": "string",
                        "example": "test1@xxx.co.jp"
                    },
                    "status": {
                        "type": "integer",
                        "example": "1"
                    },
                    "image_file": {
                        "type": "string",
                        "example": "xxxxoooo.png"
                    }
                },
                "type": "object"
            },
            "user_register": {
                "required": [
                    "id",
                    "name",
                    "email",
                    "password",
                    "status",
                    "user_agent",
                    "image_file"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "test1"
                    },
                    "password": {
                        "type": "string",
                        "example": "test1234"
                    },
                    "email": {
                        "type": "string",
                        "example": "test1@xxx.co.jp"
                    },
                    "status": {
                        "type": "integer",
                        "example": "1"
                    },
                    "user_agent": {
                        "type": "string",
                        "example": "Mozilla/5.0 (Macintosh; U; Intel Mac OS X 10_8_8) AppleWebKit/5330 (KHTML, like Gecko) Chrome/36.0.833.0 Mobile Safari/5330"
                    },
                    "image_file": {
                        "type": "string",
                        "example": "xxxxoooo.png"
                    }
                },
                "type": "object"
            }
        }
    }
}